Add files via upload
Consider this as your input string. You have to iterate it and print the type and value of the entity. If it is a data structure then go inside it and do the same. You have to keep on going nested until you reach data types.(Do this using reflect concept)